Maintain original favicon url
If you update the favicon of a tab, close the tab, and then re-open it, Chrome will initially use the updated icon,
which can be observed via `chrome.tabs.onUpdated.addListener`, and the listener is called again with the actual page favicon. Getting the favicon from the second update as the "correct" icon isn't ideal.