Not sure what "no-cors" is doing for me anymore
The comment line above doesn't explain why it's there.

I ran this to see where it was introduced:
`git log -S "no-cors" --source --all`

Didn't see anything that explained it.

My hashdrop project has a similar function that doesn't have this.

I remember having a CORS issue that it solved, but I'm no longer sure what site it was, etc. It seems reasonable that there could be a CORs issue. If I find one, I'll have to undo this commit